# apis/types/tenant_create_request.go
// Code generated by go-swagger; DO NOT EDIT.

package types

// This file was generated by the swagger tool.
// Editing this file might prove futile when you re-run the swagger generate command

import (
	"context"

	"github.com/go-openapi/strfmt"
	"github.com/go-openapi/swag"
)

// TenantCreateRequest tenant create request
//
// swagger:model TenantCreateRequest
type TenantCreateRequest struct {

	// Tenant name.
	//
	Name string `json:"name,omitempty"`

	// service
	Service string `json:"service,omitempty"`

	// storage
	Storage string `json:"storage,omitempty"`
}

// Validate validates this tenant create request
func (m *TenantCreateRequest) Validate(formats strfmt.Registry) error {
	return nil
}

// ContextValidate validates this tenant create request based on context it is used
func (m *TenantCreateRequest) ContextValidate(ctx context.Context, formats strfmt.Registry) error {
	return nil
}

// MarshalBinary interface implementation
func (m *TenantCreateRequest) MarshalBinary() ([]byte, error) {
	if m == nil {
		return nil, nil
	}
	return swag.WriteJSON(m)
}

// UnmarshalBinary interface implementation
func (m *TenantCreateRequest) UnmarshalBinary(b []byte) error {
	var res TenantCreateRequest
	if err := swag.ReadJSON(b, &res); err != nil {
		return err
	}
	*m = res
	return nil
}