# apis/types/tenant_info.go
// Code generated by go-swagger; DO NOT EDIT.

package types

// This file was generated by the swagger tool.
// Editing this file might prove futile when you re-run the swagger generate command

import (
	"context"

	"github.com/go-openapi/strfmt"
	"github.com/go-openapi/swag"
)

// TenantInfo detailed information about tenant in whizard.
//
// swagger:model TenantInfo
type TenantInfo struct {

	// compactor
	Compactor string `json:"compactor,omitempty"`

	// ingester
	Ingester string `json:"ingester,omitempty"`

	// ruler
	Ruler string `json:"ruler,omitempty"`

	// service
	Service string `json:"service,omitempty"`

	// storage
	Storage string `json:"storage,omitempty"`

	// ID of the Tenant.
	TenantID string `json:"tenantID,omitempty"`
}

// Validate validates this tenant info
func (m *TenantInfo) Validate(formats strfmt.Registry) error {
	return nil
}

// ContextValidate validates this tenant info based on context it is used
func (m *TenantInfo) ContextValidate(ctx context.Context, formats strfmt.Registry) error {
	return nil
}

// MarshalBinary interface implementation
func (m *TenantInfo) MarshalBinary() ([]byte, error) {
	if m == nil {
		return nil, nil
	}
	return swag.WriteJSON(m)
}

// UnmarshalBinary interface implementation
func (m *TenantInfo) UnmarshalBinary(b []byte) error {
	var res TenantInfo
	if err := swag.ReadJSON(b, &res); err != nil {
		return err
	}
	*m = res
	return nil
}