# apis/types/tenant_update_request.go
// Code generated by go-swagger; DO NOT EDIT.

package types

// This file was generated by the swagger tool.
// Editing this file might prove futile when you re-run the swagger generate command

import (
	"context"

	"github.com/go-openapi/strfmt"
	"github.com/go-openapi/swag"
)

// TenantUpdateRequest request used to update Tenant attributes.
//
// swagger:model TenantUpdateRequest
type TenantUpdateRequest struct {

	// service
	Service string `json:"service,omitempty"`

	// storage
	Storage string `json:"storage,omitempty"`
}

// Validate validates this tenant update request
func (m *TenantUpdateRequest) Validate(formats strfmt.Registry) error {
	return nil
}

// ContextValidate validates this tenant update request based on context it is used
func (m *TenantUpdateRequest) ContextValidate(ctx context.Context, formats strfmt.Registry) error {
	return nil
}

// MarshalBinary interface implementation
func (m *TenantUpdateRequest) MarshalBinary() ([]byte, error) {
	if m == nil {
		return nil, nil
	}
	return swag.WriteJSON(m)
}

// UnmarshalBinary interface implementation
func (m *TenantUpdateRequest) UnmarshalBinary(b []byte) error {
	var res TenantUpdateRequest
	if err := swag.ReadJSON(b, &res); err != nil {
		return err
	}
	*m = res
	return nil
}